Update Redis sink to stream sink
